+++ lc-lib/harvester/linereader_test.go
@@ -18,50 +18,29 @@ package harvester
 
 import (
 	"bytes"
+	"io"
 	"testing"
 )
 
-func checkLine(t *testing.T, reader *LineReader, expected []byte) {
+func checkLine(t *testing.T, reader *LineReader, expected []byte, expectedErr error) {
 	line, err := reader.ReadSlice()
-	if line == nil {
-		t.Log("No line returned")
-		t.FailNow()
+	if line != nil || expected != nil {
+		if line == nil {
+			t.Error("No line returned")
+		} else if expected == nil {
+			t.Errorf("Line data was not expected: [% X]", line)
+		} else if !bytes.Equal(line, expected) {
+			t.Errorf("Line data incorrect: [% X]", line)
+		}
 	}
-	if !bytes.Equal(line, expected) {
-		t.Logf("Line data incorrect: [% X]", line)
-		t.FailNow()
-	}
-	if err != nil {
-		t.Logf("Unexpected error: %s", err)
-		t.FailNow()
+	if err != expectedErr {
+		t.Errorf("Unexpected error: %s", err)
 	}
 }
 
-func checkLineTooLong(t *testing.T, reader *LineReader, expected []byte) {
-	line, err := reader.ReadSlice()
-	if line == nil {
-		t.Log("No line returned")
-		t.FailNow()
-	}
-	if !bytes.Equal(line, expected) {
-		t.Logf("Line data incorrect: [% X]", line)
-		t.FailNow()
-	}
-	if err != ErrLineTooLong {
-		t.Logf("Unexpected error: %s", err)
-		t.FailNow()
-	}
-}
-
-func checkEnd(t *testing.T, reader *LineReader) {
-	line, err := reader.ReadSlice()
-	if line != nil {
-		t.Log("Unexpected line returned")
-		t.FailNow()
-	}
-	if err == nil {
-		t.Log("Expected error")
-		t.FailNow()
+func checkBufferedLen(t *testing.T, reader *LineReader, expected int) {
+	if reader.BufferedLen() != expected {
+		t.Errorf("Incorrected buffered length: found %d != expected %d", reader.BufferedLen(), expected)
 	}
 }
 
@@ -71,9 +50,10 @@ func TestLineRead(t *testing.T) {
 	// New line read with 100 bytes, enough for the above
 	reader := NewLineReader(data, 100, 100)
 
-	checkLine(t, reader, []byte("12345678901234567890\n"))
-	checkLine(t, reader, []byte("12345678901234567890\n"))
-	checkEnd(t, reader)
+	checkLine(t, reader, []byte("12345678901234567890\n"), nil)
+	checkLine(t, reader, []byte("12345678901234567890\n"), nil)
+	checkLine(t, reader, nil, io.EOF)
+	checkBufferedLen(t, reader, 0)
 }
 
 func TestLineReadEmpty(t *testing.T) {
@@ -82,9 +62,10 @@ func TestLineReadEmpty(t *testing.T) {
 	// New line read with 100 bytes, enough for the above
 	reader := NewLineReader(data, 100, 100)
 
-	checkLine(t, reader, []byte("\n"))
-	checkLine(t, reader, []byte("12345678901234567890\n"))
-	checkEnd(t, reader)
+	checkLine(t, reader, []byte("\n"), nil)
+	checkLine(t, reader, []byte("12345678901234567890\n"), nil)
+	checkLine(t, reader, nil, io.EOF)
+	checkBufferedLen(t, reader, 0)
 }
 
 func TestLineReadIncomplete(t *testing.T) {
@@ -93,9 +74,10 @@ func TestLineReadIncomplete(t *testing.T) {
 	// New line read with 100 bytes, enough for the above
 	reader := NewLineReader(data, 100, 100)
 
-	checkLine(t, reader, []byte("\n"))
-	checkLine(t, reader, []byte("12345678901234567890\n"))
-	checkEnd(t, reader)
+	checkLine(t, reader, []byte("\n"), nil)
+	checkLine(t, reader, []byte("12345678901234567890\n"), nil)
+	checkLine(t, reader, nil, io.EOF)
+	checkBufferedLen(t, reader, 6)
 }
 
 func TestLineReadOverflow(t *testing.T) {
@@ -104,23 +86,28 @@ func TestLineReadOverflow(t *testing.T) {
 	// New line read with 21 bytes buffer but 100 max line to trigger overflow
 	reader := NewLineReader(data, 21, 100)
 
-	checkLine(t, reader, []byte("12345678901234567890\n"))
-	checkLine(t, reader, []byte("123456789012345678901234567890\n"))
-	checkLine(t, reader, []byte("12345678901234567890\n"))
-	checkEnd(t, reader)
+	checkLine(t, reader, []byte("12345678901234567890\n"), nil)
+	checkLine(t, reader, []byte("123456789012345678901234567890\n"), nil)
+	checkLine(t, reader, []byte("12345678901234567890\n"), nil)
+	checkLine(t, reader, nil, io.EOF)
+	checkBufferedLen(t, reader, 0)
 }
 
 func TestLineReadOverflowTooLong(t *testing.T) {
 	data := bytes.NewBufferString("12345678901234567890\n123456789012345678901234567890\n12345678901234567890\n")
 
 	// New line read with 10 bytes buffer and 21 max line length
+	// This test regression tests a really old bug where when a too long line
+	// was received that overflowed in a way specific to these values it would
+	// corrupt the entry
 	reader := NewLineReader(data, 10, 21)
 
-	checkLine(t, reader, []byte("12345678901234567890\n"))
-	checkLineTooLong(t, reader, []byte("123456789012345678901"))
-	checkLine(t, reader, []byte("234567890\n"))
-	checkLine(t, reader, []byte("12345678901234567890\n"))
-	checkEnd(t, reader)
+	checkLine(t, reader, []byte("12345678901234567890\n"), nil)
+	checkLine(t, reader, []byte("123456789012345678901"), ErrLineTooLong)
+	checkLine(t, reader, []byte("234567890\n"), nil)
+	checkLine(t, reader, []byte("12345678901234567890\n"), nil)
+	checkLine(t, reader, nil, io.EOF)
+	checkBufferedLen(t, reader, 0)
 }
 
 func TestLineReadTooLong(t *testing.T) {
@@ -129,9 +116,10 @@ func TestLineReadTooLong(t *testing.T) {
 	// New line read with ample buffer and 21 max line length
 	reader := NewLineReader(data, 100, 21)
 
-	checkLine(t, reader, []byte("12345678901234567890\n"))
-	checkLineTooLong(t, reader, []byte("123456789012345678901"))
-	checkLine(t, reader, []byte("234567890\n"))
-	checkLine(t, reader, []byte("12345678901234567890\n"))
-	checkEnd(t, reader)
+	checkLine(t, reader, []byte("12345678901234567890\n"), nil)
+	checkLine(t, reader, []byte("123456789012345678901"), ErrLineTooLong)
+	checkLine(t, reader, []byte("234567890\n"), nil)
+	checkLine(t, reader, []byte("12345678901234567890\n"), nil)
+	checkLine(t, reader, nil, io.EOF)
+	checkBufferedLen(t, reader, 0)
 }